注册实现

+++ microBlog/src/main/java/com/microblog/bean/User.java
@@ -10,7 +10,7 @@ public class User implements Serializable{
 	private String password;
 	private String nickname;
 	private String pic;
-	private Integer telephone;
+	private Long telephone;
 	private Integer level;
 	private Integer exp;
 	private Date regDate;
@@ -48,10 +48,10 @@ public class User implements Serializable{
 	public void setPic(String pic) {
 		this.pic = pic;
 	}
-	public Integer getTelephone() {
+	public Long getTelephone() {
 		return telephone;
 	}
-	public void setTelephone(Integer telephone) {
+	public void setTelephone(Long telephone) {
 		this.telephone = telephone;
 	}
 	public Integer getLevel() {
